fix bug

+++ src/ClientNode.java
@@ -1,6 +1,6 @@
 public class ClientNode extends Point {
     int demand;
-    public ClientNode(int id, float x, float y, int demand) {
+    public ClientNode(int id, double x, double y, int demand) {
         super(id, x, y);
         this.demand = demand;
     }
