+++ src/DepotNode.java
@@ -1,6 +1,6 @@
 public class DepotNode extends Point{
 
-    public DepotNode(int id, float x, float y) {
+    public DepotNode(int id, double x, double y) {
         super(id, x, y);
     }
 }
