+++ src/PointEvaluationRes.java
@@ -2,10 +2,10 @@
 
 public class PointEvaluationRes implements Comparable<PointEvaluationRes> {
     public Point point;
-    public float score;
+    public double score;
 
     public Route route;
-    public PointEvaluationRes(Route route,Point point, float score){
+    public PointEvaluationRes(Route route, Point point, double score){
         this.route = route;
         this.point = point;
         this.score = score;
@@ -13,6 +13,6 @@ public class PointEvaluationRes implements Comparable<PointEvaluationRes> {
 
     @Override
     public int compareTo(PointEvaluationRes o) {
-        return Float.compare(score, o.score);
+        return Double.compare(score, o.score);
     }
 }
